rename Archivable to PathRestorable. move NavigationLink extensions into PathRestorable file
